Rename validator package to content, move soc validator to soc
